update regex for removing more characters from team name update image upload component to accept being embedded as a parameter to add some admin options add mephisto to heroes ban list add admin routes and service handlers for admin team logo upload and removal refactor the aws image upload code into a reusable method
