+++ client/src/app/team-profile/team-profile.component.html
@@ -96,7 +96,7 @@
 
                 <div class="row mt-3">
                     <div class="col-sm-12 col-md-5">
-                        <app-image-upload [showEdit]="auth.getUser()===returnedProfile.captain" [teamName]="returnedProfile.teamName" [teamLogo]="returnedProfile.logo"></app-image-upload>
+                        <app-image-upload [showEdit]="auth.getUser()===returnedProfile.captain || componentEmbedded" [embedded]="componentEmbedded" [teamName]="returnedProfile.teamName" [teamLogo]="returnedProfile.logo"></app-image-upload>
                     </div>
                     <div class="col-sm-12 col-md-7">
                         <mat-form-field class="mat-FullWidth" *ngIf="componentEmbedded">
